requirements.txt
Flask==1.1.2
requests==2.25.1
responses==0.13.3
pytest==6.2.2
PyYAML==5.4.1